rename install to apply
