+++ examples/multipass/README.md
@@ -6,5 +6,5 @@ This directory provides an example flow with mcc tool together with [Multipass](
 
 1. Create machines `multipass launch -n manager --mem 4G --disk 10G && multipass launch -n worker --disk 10G`
 2. Update IP addresses in `cluster.yml` (run `multipass ls` to find correct IPs)
-3. `sudo ./mcc install -c ./examples/multipass/cluster.yml` (access to ssh key file requires sudo)
+3. `sudo launchpad apply -c ./examples/multipass/cluster.yml` (access to ssh key file requires sudo)
 4. Profit! :)
