+++ examples/tf-hetzner/README.md
@@ -8,6 +8,6 @@ This directory provides an example flow with mcc tool together with Terraform.
 2. `terraform init`
 3. `terraform apply`
 4. `terraform output -json | yq r --prettyPrint - ucp_cluster.value > cluster.yaml `
-5. `mcc install --config ./examples/tf-hetzner/cluster.yaml`
+5. `launchpad apply --config ./examples/tf-hetzner/cluster.yaml`
 6. Profit! :)
