+++ pkg/cmd/register/register.go
@@ -8,45 +8,38 @@ import (
 	"github.com/Mirantis/mcc/pkg/analytics"
 	"github.com/Mirantis/mcc/pkg/config"
 	log "github.com/sirupsen/logrus"
-	"github.com/urfave/cli/v2"
 )
 
 // Register ...
-func Register(ctx *cli.Context) error {
-	userConfig := config.UserConfig{
-		Name:    ctx.String("name"),
-		Company: ctx.String("company"),
-		Email:   ctx.String("email"),
-	}
-
+func Register(userConfig *config.UserConfig) error {
 	icons := func(icons *survey.IconSet) {
 		icons.Question.Text = ">"
 	}
 
 	if validateName(userConfig.Name) != nil {
-		err := survey.AskOne(&survey.Input{Message: "Name"}, &userConfig.Name, survey.WithValidator(validateName), survey.WithIcons(icons))
+		err := survey.AskOne(&survey.Input{Message: "Name"}, userConfig.Name, survey.WithValidator(validateName), survey.WithIcons(icons))
 		if err != nil {
 			return err
 		}
 	}
 
 	if validateEmail(userConfig.Email) != nil {
-		err := survey.AskOne(&survey.Input{Message: "Email"}, &userConfig.Email, survey.WithValidator(validateEmail), survey.WithIcons(icons))
+		err := survey.AskOne(&survey.Input{Message: "Email"}, userConfig.Email, survey.WithValidator(validateEmail), survey.WithIcons(icons))
 		if err != nil {
 			return err
 		}
 	}
 
 	if userConfig.Company == "" {
-		err := survey.AskOne(&survey.Input{Message: "Company"}, &userConfig.Company, survey.WithIcons(icons))
+		err := survey.AskOne(&survey.Input{Message: "Company"}, userConfig.Company, survey.WithIcons(icons))
 		if err != nil {
 			return err
 		}
 	}
 
-	err := config.SaveUserConfig(&userConfig)
+	err := config.SaveUserConfig(userConfig)
 	if err == nil {
-		analytics.IdentifyUser(&userConfig)
+		analytics.IdentifyUser(userConfig)
 		log.Info("Registration completed!")
 	} else {
 		log.Error("Registration failed!")
